Move some funcs/classes into a utils module
